CLI and GUI
text_editor_undo_redo.c is a CLI demo on stack demonstrating LIFO operations and the GUI does the same rather the instead of having save points like pressing enter in the CLI to go for another operation, you must click on commit so that the undo can work well other wise all the text will be stored in one buffer and the operation will occur on the whole block.

More adjustments are underway.